Add files via upload
This is my first upload. I was working on making the background stipes responsive, which I think I mainly did, but I don't think this will be the best way to code it because then the content on top of the stipes that line up would also have to be responsive. I'm not sure. I also don't know why my hero image doesn't go all the way across the page at larger screen sizes.